LMN-1303 Remove focus from text input when scrolling in BPKAppSearchModal set autocorrection disabled = true for BPKAppSearchModal
